fix: move ensadmin-specific util into ensadmin and out of datasources
